Adding input string normalizer.
this is getting to be rather hacky, but let's go with this for now.

i hit a lot of problems with F# lexer where i found it hard to escape certain characters from string literals. to get around this qp part had to be sacrifised and now new class called InputStringNormalizer is introduced.

Callers are responsible to push every Sql.value of string type through this class so replacement tokens can be applied.

This is all very ugly, but test coverage should help with future problems.

in future, it is likely  that we will get away from f# yacc, so this should go away with that.

it is unfortunate that problems in yacc had to spill to qp part...